Switched fiducial detection from f64 to f32. The probing is 10x faster, giving an overall 13hz.
totalTime: 75ms
Event DetectFiducialMarkers: Mean:75356us Min:75356us Max:75356us Total:75356us NumEvents:1
Event ExtractComponentsViaCharacteristicScale: Mean:62316us Min:62316us Max:62316us Total:62316us NumEvents:1
Event ccsiabaec_init: Mean:4173us Min:4173us Max:4173us Total:4173us NumEvents:1
Event ccsiabaec_mainLoop: Mean:57166us Min:57166us Max:57166us Total:57166us NumEvents:1
Event ccsiabaec_filterRows: Mean:103us Min:103us Max:104us Total:24768us NumEvents:240
Event ccsiabaec_computeBinaryImage: Mean:85us Min:84us Max:85us Total:20307us NumEvents:240
Event ccsiabaec_extractNextRowOfComponents: Mean:17us Min:14us Max:26us Total:4067us NumEvents:240
Event ccsiabaec_scrollIntegralImage: Mean:32us Min:0us Max:1096us Total:7594us NumEvents:240
Event ccsiabaec_finalize: Mean:76us Min:76us Max:76us Total:76us NumEvents:1
Event CompressConnectedComponentSegmentIds1: Mean:61us Min:61us Max:61us Total:61us NumEvents:1
Event InvalidateSmallOrLargeComponents: Mean:76us Min:76us Max:76us Total:76us NumEvents:1
Event CompressConnectedComponentSegmentIds2: Mean:57us Min:57us Max:57us Total:57us NumEvents:1
Event InvalidateSolidOrSparseComponents: Mean:133us Min:133us Max:133us Total:133us NumEvents:1
Event CompressConnectedComponentSegmentIds3: Mean:53us Min:53us Max:53us Total:53us NumEvents:1
Event InvalidateFilledCenterComponents: Mean:147us Min:147us Max:147us Total:147us NumEvents:1
Event CompressConnectedComponentSegmentIds4: Mean:53us Min:53us Max:53us Total:53us NumEvents:1
Event SortConnectedComponentSegmentsById: Mean:72us Min:72us Max:72us Total:72us NumEvents:1
Event ComputeQuadrilateralsFromConnectedComponents: Mean:2466us Min:2466us Max:2466us Total:2466us NumEvents:1
Event ComputeHomographyFromQuad: Mean:123us Min:123us Max:123us Total:123us NumEvents:1
Event Copy Image To Onchip: Mean:6138us Min:6138us Max:6138us Total:6138us NumEvents:1
Event ExtractBlockMarker: Mean:2735us Min:2735us Max:2735us Total:2735us NumEvents:1
Event fmpebm.init: Mean:6us Min:6us Max:6us Total:6us NumEvents:1
Event fmpebm.extractMean: Mean:2690us Min:2690us Max:2690us Total:2690us NumEvents:1
Event fmpebm.orient: Mean:14us Min:14us Max:14us Total:14us NumEvents:1
Event fmpebm.decode: Mean:20us Min:20us Max:20us Total:20us NumEvents:1

+++ coretech/vision/include/anki/vision/robot/fiducialDetection.h
@@ -23,7 +23,7 @@ namespace Anki
     Result DetectFiducialMarkers(
       const Array<u8> &image,
       FixedLengthList<BlockMarker> &markers,
-      FixedLengthList<Array<f64> > &homographies,
+      FixedLengthList<Array<f32> > &homographies,
       const s32 scaleImage_numPyramidLevels, const s32 scaleImage_thresholdMultiplier,
       const s16 component1d_minComponentWidth, const s16 component1d_maxSkipDistance,
       const s32 component_minimumNumPixels, const s32 component_maximumNumPixels,
