+++ coretech/vision/robot/src/transformations.cpp
@@ -459,7 +459,7 @@ namespace Anki
         return RESULT_OK;
       }
 
-      Result ComputeHomographyFromQuad(const Quadrilateral<s16> &quad, Array<f64> &homography, MemoryStack scratch)
+      Result ComputeHomographyFromQuad(const Quadrilateral<s16> &quad, Array<f32> &homography, MemoryStack scratch)
       {
         Result lastResult;
 
@@ -469,24 +469,24 @@ namespace Anki
         AnkiConditionalErrorAndReturnValue(scratch.IsValid(),
           RESULT_FAIL_INVALID_OBJECT, "ComputeHomographyFromQuad", "scratch is not valid");
 
-        FixedLengthList<Point<f64> > originalPoints(4, scratch);
-        FixedLengthList<Point<f64> > transformedPoints(4, scratch);
+        FixedLengthList<Point<f32> > originalPoints(4, scratch);
+        FixedLengthList<Point<f32> > transformedPoints(4, scratch);
 
-        originalPoints.PushBack(Point<f64>(0,0));
-        originalPoints.PushBack(Point<f64>(0,1));
-        originalPoints.PushBack(Point<f64>(1,0));
-        originalPoints.PushBack(Point<f64>(1,1));
+        originalPoints.PushBack(Point<f32>(0,0));
+        originalPoints.PushBack(Point<f32>(0,1));
+        originalPoints.PushBack(Point<f32>(1,0));
+        originalPoints.PushBack(Point<f32>(1,1));
 
-        transformedPoints.PushBack(Point<f64>(quad[0].x, quad[0].y));
-        transformedPoints.PushBack(Point<f64>(quad[1].x, quad[1].y));
-        transformedPoints.PushBack(Point<f64>(quad[2].x, quad[2].y));
-        transformedPoints.PushBack(Point<f64>(quad[3].x, quad[3].y));
+        transformedPoints.PushBack(Point<f32>(quad[0].x, quad[0].y));
+        transformedPoints.PushBack(Point<f32>(quad[1].x, quad[1].y));
+        transformedPoints.PushBack(Point<f32>(quad[2].x, quad[2].y));
+        transformedPoints.PushBack(Point<f32>(quad[3].x, quad[3].y));
 
         if((lastResult = Matrix::EstimateHomography(originalPoints, transformedPoints, homography, scratch)) != RESULT_OK)
           return lastResult;
 
         return RESULT_OK;
-      } // Result ComputeHomographyFromQuad(FixedLengthList<Quadrilateral<s16> > quads, FixedLengthList<Array<f64> > &homographies, MemoryStack scratch)
+      } // Result ComputeHomographyFromQuad(FixedLengthList<Quadrilateral<s16> > quads, FixedLengthList<Array<f32> > &homographies, MemoryStack scratch)
     } // namespace Transformations
   } // namespace Embedded
 } // namespace Anki
